pyproject.toml
[tool.poetry]
name = "redditboersenbrief"
version = "0.1.0"
description = ""
authors = ["Sören Michaels <soeren.michaels@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
requests = "^2.25.1"
praw = "^7.2.0"
pypandoc = "^1.5"
urlextract = "^1.2.0"
python-dotenv = "^0.17.1"
Pillow = "^8.2.0"
pandas = "^1.2.4"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
